Fix unit tests for context picker UI
